Moved octree-related streaming logic to its own file
